+++ steinbock/export/_cli/__init__.py
@@ -56,11 +56,15 @@ def ome_cmd(img_dir, panel_file, ome_dir):
     ]
     Path(ome_dir).mkdir(exist_ok=True)
     for img_file in io.list_image_files(img_dir):
-        img = io.read_image(img_file)
+        img = io.read_image(img_file, native_dtype=True)
         ome_file = io._as_path_with_suffix(
             Path(ome_dir) / img_file.name, ".ome.tiff"
         )
-        xtiff.to_tiff(img, ome_file, channel_names=channel_names)
+        xtiff.to_tiff(
+            io._to_dtype(img, np.float32),
+            ome_file,
+            channel_names=channel_names,
+        )
         click.echo(ome_file)
         del img
 
@@ -131,7 +135,7 @@ def histocat_cmd(img_dir, mask_dir, panel_file, histocat_dir):
             click.echo(histocat_img_file)
         mask = None
         if mask_files is not None:
-            mask = io.read_mask(mask_files[i])
+            mask = io.read_mask(mask_files[i], native_dtype=True)
             histocat_mask_file = (
                 histocat_img_dir / f"{img_file.stem}_mask.tiff"
             )