+++ steinbock/utils/_cli/matching.py
@@ -14,13 +14,20 @@ from steinbock.utils import matching
     "masks2", nargs=1, type=click.Path(exists=True, file_okay=False)
 )
 @click.option(
+    "--mmap/--no-mmap",
+    "mmap",
+    default=False,
+    show_default=True,
+    help="Use memory mapping for reading images/masks",
+)
+@click.option(
     "-o",
     "csv_dir",
     type=click.Path(file_okay=False),
     required=True,
     help="Path to the object table CSV output directory",
 )
-def match_cmd(masks1, masks2, csv_dir):
+def match_cmd(masks1, masks2, mmap, csv_dir):
     if Path(masks1).is_file() and Path(masks2).is_file():
         mask_files1 = [Path(masks1)]
         mask_files2 = [Path(masks2)]
@@ -29,7 +36,7 @@ def match_cmd(masks1, masks2, csv_dir):
         mask_files2 = io.list_mask_files(masks2, base_files=mask_files1)
     Path(csv_dir).mkdir(exist_ok=True)
     for mask_file1, mask_file2, df in matching.try_match_masks_from_disk(
-        mask_files1, mask_files2
+        mask_files1, mask_files2, mmap=mmap
     ):
         csv_file = io._as_path_with_suffix(
             Path(csv_dir) / mask_file1.name, ".csv"