+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2022-01-26
+
+  - Full memory mapping support
+  - Fixed zero division error in deepcell normalization
+  - Added GPU support (switched to Ubuntu 18.04-based tensorflow base image)
+  - Fixed IMC panel preprocessing 
+  - Upgraded dependencies
 
 ## [0.12.0] - 2022-01-08
 
@@ -279,6 +286,7 @@ Added:
 Initial release for beta testing
 
 
+[0.13.0]: https://github.com/BodenmillerGroup/steinbock/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/BodenmillerGroup/steinbock/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/BodenmillerGroup/steinbock/compare/v0.10.7...v0.11.0
 [0.10.7]: https://github.com/BodenmillerGroup/steinbock/compare/v0.10.6...v0.10.7